more beautification

+++ CompoundFs/FileSharedMutex.h
@@ -11,9 +11,9 @@ namespace TxFs
     {
     public:
         FileSharedMutex() = default;
-        FileSharedMutex(void* handle, uint64_t begin, uint64_t end)
+        FileSharedMutex(void* handle, int64_t begin, int64_t end)
             : m_fileHandle(handle)
-            , m_lockRange(begin, end)
+            , m_lockRange((uint64_t) begin, (uint64_t) end)
         {
 
         }
